Move `ttyname` into the `termios` module.
