+++ src/imp/libc/termios/syscalls.rs
@@ -9,6 +9,9 @@
 use super::super::c;
 use super::super::conv::{borrowed_fd, ret, ret_pid_t};
 use crate::fd::BorrowedFd;
+#[cfg(feature = "procfs")]
+#[cfg(not(any(target_os = "fuchsia", target_os = "wasi")))]
+use crate::ffi::ZStr;
 use crate::io;
 use crate::process::{Pid, RawNonZeroPid};
 use crate::termios::{Action, OptionalActions, QueueSelector, Speed, Termios, Winsize};
@@ -140,5 +143,18 @@ pub(crate) fn isatty(fd: BorrowedFd<'_>) -> bool {
         }
     } else {
         true
+    }
+}
+
+#[cfg(feature = "procfs")]
+#[cfg(not(any(target_os = "fuchsia", target_os = "wasi")))]
+pub(crate) fn ttyname(dirfd: BorrowedFd<'_>, buf: &mut [u8]) -> io::Result<usize> {
+    unsafe {
+        ret(c::ttyname_r(
+            borrowed_fd(dirfd),
+            buf.as_mut_ptr().cast(),
+            buf.len(),
+        ))?;
+        Ok(ZStr::from_ptr(buf.as_ptr().cast()).to_bytes().len())
     }
 }
