+++ tests/termios/main.rs
@@ -1,0 +1,8 @@
+//! Tests for [`rustix::termios`].
+
+#![cfg_attr(target_os = "wasi", feature(wasi_ext))]
+#![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
+#![cfg(feature = "termios")]
+
+#[cfg(not(windows))]
+mod isatty;
